fix give registration coupon

+++ app/Services/CouponService.php
@@ -320,7 +320,7 @@ class CouponService
                 'type'          => Coupon::getTypeIdByName('all'),
                 'name'          => trans('front_labels.registration_coupon'),
                 'description'   => trans('front_texts.registration_coupon_description'),
-                'discount'      => variable('registration_coupon_discount'),
+                'discount'      => (int) variable('registration_coupon_discount'),
                 'discount_type' => Coupon::getDiscountTypeIdByName('percentage'),
                 'count'         => 1,
                 'users_count'   => 1,
